First part of converting tests to compile_tools
