+++ src/amuse/test/compile_tools.py
@@ -1,8 +1,11 @@
 import os
 import subprocess
 import time
+import shlex
 
 from amuse.rfi.tools import create_c
+from amuse.rfi.core import config
+from amuse.test.amusetest import get_amuse_root_dir
 
 
 def get_mpicc_name():
@@ -57,6 +60,83 @@ def get_mpicxx_flags():
         return ""
 
 
+def is_fortran_version_up_to_date():
+    try:
+        from amuse import config
+        is_configured = hasattr(config, 'compilers')
+        if is_configured:
+            is_configured = hasattr(config.compilers, 'gfortran_version')
+    except ImportError:
+        is_configured = False
+
+    if is_configured:
+        if not config.compilers.gfortran_version:
+            if not hasattr(config.compilers, 'ifort_version') or not config.compilers.ifort_version:
+                return True
+            try:
+                parts = [int(x) for x in config.compilers.ifort_version.split('.')]
+            except:
+                parts = []
+
+            return parts[0] > 9  
+
+        try:
+            parts = [int(x) for x in config.compilers.gfortran_version.split('.')]
+        except:
+            parts = []
+
+        if len(parts) < 2:
+            return True
+
+        return parts[0] >= 4 and parts[1] >= 3
+    else:
+        return True
+
+
+def get_mpif90_name():
+    try:
+        from amuse import config
+        is_configured = hasattr(config, 'mpi')
+    except ImportError:
+        is_configured = False
+
+    if is_configured:
+        return config.mpi.mpif95
+    else:
+        return os.environ['MPIFC'] if 'MPIFC' in os.environ else 'mpif90'
+
+
+def get_mpif90_arguments():
+    name = get_mpif90_name()
+    return list(shlex.split(name))
+
+
+def has_fortran_iso_c_binding():
+    try:
+        from amuse import config
+        is_configured = hasattr(config, 'mpi')
+    except ImportError:
+        is_configured = False
+
+    if is_configured:
+        return config.compilers.fc_iso_c_bindings
+    else:
+        return False
+
+
+def get_ld_flags():
+    try:
+        from amuse import config
+        is_configured = hasattr(config, 'compilers')
+    except ImportError:
+        is_configured = False
+
+    if is_configured:
+        return config.compilers.ld_flags
+    else:
+        return ""
+
+
 def wait_for_file(filename):
     for dt in [0.01, 0.01, 0.02, 0.05]:
         if os.path.exists(filename):
@@ -64,7 +144,18 @@ def wait_for_file(filename):
         time.sleep(dt)
 
 
-def c_compile(objectname, string):
+def open_subprocess(arguments):
+    process = subprocess.Popen(
+        arguments,
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE
+    )
+    stdout, stderr = process.communicate()
+    return process, stdout, stderr
+
+
+def c_compile(objectname, string, extra_args=[]):
     root, ext = os.path.splitext(objectname)
     sourcename = root + '.c'
 
@@ -77,16 +168,10 @@ def c_compile(objectname, string):
     mpicc = get_mpicc_name()
     arguments = [mpicc]
     arguments.extend(get_mpicc_flags().split())
-    arguments.extend(["-I", "lib/stopcond", "-c",  "-o", objectname,
-                     sourcename])
+    arguments.extend(["-I", "lib/stopcond", "-c"] + extra_args + ["-o",
+                     objectname, sourcename])
 
-    process = subprocess.Popen(
-        arguments,
-        stdin=subprocess.PIPE,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-    )
-    stdout, stderr = process.communicate()
+    process, stderr, stdout = open_subprocess(arguments)
 
     if process.returncode == 0:
         wait_for_file(objectname)
@@ -113,13 +198,7 @@ def cxx_compile(objectname, string):
     arguments.extend(["-I", "lib/stopcond", "-c",  "-o", objectname,
                      sourcename])
 
-    process = subprocess.Popen(
-        arguments,
-        stdin=subprocess.PIPE,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-    )
-    stdout, stderr = process.communicate()
+    process, stderr, stdout = open_subprocess(arguments)
 
     if process.returncode == 0:
         wait_for_file(objectname)
@@ -147,13 +226,9 @@ def c_build(exename, objectnames):
         libs = os.environ['LIBS'].split()
         arguments.extend(libs)
 
-    process = subprocess.Popen(
-        arguments,
-        stdin=subprocess.PIPE,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE
-    )
-    stdout, stderr = process.communicate()
+    arguments.extend(get_ld_flags().split())
+
+    process, stderr, stdout = open_subprocess(arguments)
 
     if process.returncode == 0:
         wait_for_file(exename)
@@ -192,3 +267,189 @@ def build_worker(codestring, path_to_results, specification_class):
 
     cxx_compile(interfacefile, code)
     c_build(exefile, [interfacefile, codefile])
+
+    return exefile
+
+
+def c_pythondev_compile(objectname, string):
+    root, ext = os.path.splitext(objectname)
+    sourcename = root + '.c'
+    amuse_root = get_amuse_root_dir()
+    if os.path.exists(objectname):
+        os.remove(objectname)
+
+    with open(sourcename, "w") as f:
+        f.write(string)
+
+    mpicc = get_mpicc_name()
+    arguments = [mpicc]
+    arguments.extend(get_mpicc_flags().split())
+
+    arguments.extend(["-I", amuse_root + "/lib/stopcond","-I", amuse_root + "/lib/amuse_mpi",  "-fPIC", "-c",  "-o", objectname, sourcename])
+    arguments.extend(shlex.split(config.compilers.pythondev_cflags))
+
+    process, stderr, stdout = open_subprocess(arguments)
+
+    if process.returncode == 0:
+        wait_for_file(objectname)
+
+    if process.returncode != 0 or not os.path.exists(objectname):
+        print "Could not compile {0}, error = {1}".format(objectname, stderr)
+        raise Exception("Could not compile {0}, error = {1}".format(objectname, stderr))
+
+
+def c_pythondev_build(self, exename, objectnames):
+    if os.path.exists(exename):
+        os.remove(exename)
+
+    mpicxx = get_mpicxx_name()
+    arguments = [mpicxx]
+    arguments.extend(objectnames)
+    arguments.extend(shlex.split(config.compilers.pythondev_ldflags))
+    arguments.append("-o")
+    arguments.append(exename)
+
+    if 'LIBS' in os.environ:
+        libs = os.environ['LIBS'].split()
+        arguments.extend(libs)
+
+    arguments.extend(get_ld_flags().split())
+
+    process, stderr, stdout = open_subprocess(arguments)
+
+    if process.returncode == 0:
+        wait_for_file(exename)
+
+    if process.returncode != 0 or not (os.path.exists(exename) or os.path.exists(exename+'.exe')):
+        print "Could not compile {0}, error = {1}".format(exename, stderr)
+        raise Exception("Could not build {0}, error = {1}".format(exename, stderr))
+
+    print stdout
+    print stderr
+
+
+def c_pythondev_buildso(soname, objectnames):
+    if os.path.exists(soname):
+        os.remove(soname)
+    amuse_root = get_amuse_root_dir()
+
+    mpicxx = get_mpicxx_name()
+    arguments = [mpicxx]
+    arguments.extend(objectnames)
+    arguments.extend(shlex.split(config.compilers.pythondev_ldflags))
+    arguments.append("-shared")
+    arguments.append("-o")
+    arguments.append(soname)
+    arguments.append("-L"+amuse_root+"/lib/amuse_mpi")
+    arguments.append("-lamuse_mpi")
+
+    if 'LIBS' in os.environ:
+        libs = os.environ['LIBS'].split()
+        arguments.extend(libs)
+
+    arguments.extend(get_ld_flags().split())
+
+    process, stderr, stdout = open_subprocess(arguments)
+
+    if process.returncode == 0:
+        wait_for_file(soname)
+
+    if process.returncode != 0 or not (os.path.exists(soname)):
+        print "Could not compile {0}, error = {1}".format(soname, stderr)
+        raise Exception("Could not build {0}, error = {1}".format(soname, stderr))
+
+    print stdout
+    print stderr
+
+
+def fortran_pythondev_buildso(self, soname, objectnames):
+    if os.path.exists(soname):
+        os.remove(soname)
+    amuse_root = get_amuse_root_dir()
+
+    arguments = get_mpif90_arguments()
+    arguments.extend(objectnames)
+    arguments.extend(shlex.split(config.compilers.pythondev_ldflags))
+    arguments.append("-shared")
+    arguments.append("-o")
+    arguments.append(soname)
+    arguments.append("-L"+amuse_root+"/lib/amuse_mpi")
+    arguments.append("-lamuse_mpi")
+
+    if 'LIBS' in os.environ:
+        libs = os.environ['LIBS'].split()
+        arguments.extend(libs)
+
+    arguments.extend(get_ld_flags().split())
+
+    process, stderr, stdout = open_subprocess(arguments)
+
+    if process.returncode == 0:
+        wait_for_file(soname)
+
+    if process.returncode != 0 or not (os.path.exists(soname)):
+        print "Could not compile {0}, error = {1}".format(soname, stderr)
+        raise Exception("Could not build {0}, error = {1}".format(soname, stderr))
+
+    print stdout
+    print stderr
+
+
+def fortran_compile(objectname, string, extra_args=[]):
+    if os.path.exists(objectname):
+        os.remove(objectname)
+
+    root, ext = os.path.splitext(objectname)
+    sourcename = root + '.f90'
+
+    with open(sourcename, "w") as f:
+        f.write(string)
+
+    arguments = get_mpif90_arguments()
+    arguments.extend(["-g",
+                     "-I{0}/lib/forsockets".format(get_amuse_root_dir())]
+                     + extra_args + ["-c",  "-o", objectname, sourcename])
+    arguments.append("-Wall")
+
+    process, stderr, stdout = open_subprocess(arguments)
+    process.wait()
+
+    if process.returncode == 0:
+        wait_for_file(objectname)
+
+    if process.returncode != 0 or not os.path.exists(objectname):
+        print "Could not compile {0}, error = {1}".format(objectname, stderr)
+        raise Exception("Could not compile {0}, error = {1}".format(objectname, stderr))
+
+    print stdout
+    print stderr
+
+
+def fortran_build(exename, objectnames):
+    if os.path.exists(exename):
+        os.remove(exename)
+
+    arguments = get_mpif90_arguments()
+    arguments.extend(objectnames)
+
+    arguments.append("-L{0}/lib/forsockets".format(get_amuse_root_dir()))
+    arguments.append("-Wall")
+    arguments.append("-lforsockets")
+
+    arguments.append("-o")
+    arguments.append(exename)
+
+    arguments.extend(get_ld_flags().split())
+
+    process, stderr, stdout = open_subprocess(arguments)
+
+    if process.returncode == 0:
+        wait_for_file(exename)
+
+    if process.returncode != 0 or not os.path.exists(exename):
+        print "Could not build {0}, error = {1}".format(exename, stderr)
+        raise Exception("Could not build {0}, error = {1}".format(exename,
+                                                                  stderr))
+
+    print stdout
+    print stderr
