+++ test/codes_tests/test_fortran_sockets_implementation.py
@@ -1,4 +1,5 @@
 from amuse.test.amusetest import TestWithMPI
+import amuse.test.compile_tools as compile_tools
 import subprocess
 import os
 import time
@@ -278,122 +279,9 @@ class ForTestingInterface(CodeInterface):
         function.result_type = 'int32'
         function.can_handle_array = True
         return function  
-    
-    
-    
+
+
 class TestInterface(TestWithMPI):
-    
-    def is_fortan_version_up_to_date(self):
-        try:
-            from amuse import config
-            is_configured = hasattr(config, 'compilers')
-            if is_configured:
-                is_configured = hasattr(config.compilers, 'gfortran_version')
-        except ImportError:
-            is_configured = False
-    
-        if is_configured:
-            if not config.compilers.gfortran_version:
-                if not hasattr(config.compilers, 'ifort_version') or not config.compilers.ifort_version:
-                    return True
-                try:
-                    parts = [int(x) for x in config.compilers.ifort_version.split('.')]
-                except:
-                    parts = []
-                    
-                return parts[0] > 9  
-                
-            try:
-                parts = [int(x) for x in config.compilers.gfortran_version.split('.')]
-            except:
-                parts = []
-                
-            if len(parts) < 2:
-                return True
-                
-            return parts[0] >= 4 and parts[1] >= 3
-        else:
-            return True
-            
-    def get_mpif90_name(self):
-        try:
-            from amuse import config
-            is_configured = hasattr(config, 'mpi')
-        except ImportError:
-            is_configured = False
-    
-        if is_configured:
-            return config.mpi.mpif95
-        else:
-            return os.environ['MPIFC'] if 'MPIFC' in os.environ else 'mpif90'
-            
-    def get_mpif90_arguments(self):
-        name = self.get_mpif90_name()
-        return list(shlex.split(name))
-    
-    def wait_for_file(self, filename):
-        for dt in [0.01, 0.01, 0.02, 0.05]:
-            if os.path.exists(filename):
-                return
-            time.sleep(dt)
-            
-    def fortran_compile(self, objectname, string):
-        if os.path.exists(objectname):
-            os.remove(objectname)
-        
-        root, ext = os.path.splitext(objectname)
-        sourcename = root + '.f90'
-        
-        with open(sourcename, "w") as f:
-            f.write(string)
-        
-            
-        arguments = self.get_mpif90_arguments()
-        arguments.extend(["-g", "-I{0}/lib/forsockets".format(self.get_amuse_root_dir()), "-c",  "-o" , objectname, sourcename])
-        
-        process = subprocess.Popen(
-            arguments, 
-            stdin = subprocess.PIPE,
-            stdout = subprocess.PIPE,
-            stderr = subprocess.PIPE
-        )
-        stdout, stderr = process.communicate()
-        process.wait()
-        
-        if process.returncode == 0:
-            self.wait_for_file(objectname)
-            
-        if process.returncode != 0 or not os.path.exists(objectname):
-            print "Could not compile {0}, error = {1}".format(objectname, stderr)
-            raise Exception("Could not compile {0}, error = {1}".format(objectname, stderr))
-            
-    
-    def fortran_build(self, exename, objectnames):
-        if os.path.exists(exename):
-            os.remove(exename)
-            
-        arguments = self.get_mpif90_arguments()
-        arguments.extend(objectnames)
-        arguments.append("-o")
-        arguments.append(exename)
-        arguments.append("-L{0}/lib/forsockets".format(self.get_amuse_root_dir()))
-        arguments.append("-lforsockets")
-        
-        process = subprocess.Popen(
-            arguments,
-            stdin = subprocess.PIPE,
-            stdout = subprocess.PIPE,
-            stderr = subprocess.PIPE
-        )
-        stdout, stderr = process.communicate()
-        
-        if process.returncode == 0:
-            self.wait_for_file(exename)
-            
-        
-        if process.returncode != 0 or not os.path.exists(exename):
-            print "Could not build {0}, error = {1}".format(exename, stderr)
-            raise Exception("Could not build {0}, error = {1}".format(exename, stderr))
     
     def build_worker(self):
         
@@ -402,13 +290,13 @@ class TestInterface(TestWithMPI):
         interfacefile = os.path.join(path,"interface-sockets.o")
         self.exefile = os.path.join(path,"fortran_worker")
         
-        self.fortran_compile(codefile, codestring)
+        compile_tools.fortran_compile(codefile, codestring)
         
         uc = create_fortran.GenerateAFortranSourcecodeStringFromASpecificationClass()
         uc.specification_class = ForTestingInterface
         string =  uc.result
-        self.fortran_compile(interfacefile, string)
-        self.fortran_build(self.exefile, [interfacefile, codefile] )
+        compile_tools.fortran_compile(interfacefile, string)
+        compile_tools.fortran_build(self.exefile, [interfacefile, codefile] )
     
     def setUp(self):
         super(TestInterface, self).setUp()
@@ -419,7 +307,7 @@ class TestInterface(TestWithMPI):
         self.build_worker()
         
     def check_fortran_version(self):
-        if not self.is_fortan_version_up_to_date():
+        if not compile_tools.is_fortran_version_up_to_date():
             self.skip('cannot compile fortran socket modules with old fortran compilers (missing C support)')
         
     def check_not_in_mpiexec(self):
